# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "是否需要分页， 默认需要， 如果不分页 传 true",
                        "name": "isNotPaging",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页码， 如果不分页 传 0",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "一页大小， 如果不分页 传 0",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务描述",
                        "name": "mark",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Full"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ReqQueryUser"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "description": "修改用户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqModifyUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "description": "添加用户",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqCreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
        "/api/user/status/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "停用/启用用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "停用/启用用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.ReqCreateUser": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "登录名",
                    "type": "string"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "用户名",
                    "type": "string"
                },
                "password": {
                    "description": "用户密码",
                    "type": "string"
                }
            }
        },
...
        "model.ReqModifyUser": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "登录名",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "用户名",
                    "type": "string"
                },
                "password": {
                    "description": "用户密码",
                    "type": "string"
                }
            }
        },
...
        "model.ReqQueryUser": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "登录名",
                    "type": "string"
                },
                "isNotPaging": {
                    "description": "是否分页",
                    "type": "boolean"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "用户名",
                    "type": "string"
                },
                "pageNum": {
                    "description": "页数",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "页码",
                    "type": "integer"
                },
                "total": {
                    "description": "总数  由服务器返回回去",
                    "type": "integer"
                }
            }
        },
...
        "res.Base": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/res.RCode"
                        }
                    ]
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "res.Full": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/res.RCode"
                        }
                    ]
                },
                "data": {
                    "description": "返回数据"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "res.RCode": {
            "type": "integer",
            "enum": [
                200,
                201,
                400,
                401,
                404,
                405,
                500,
                501,
                502,
                504,
                505
            ],
            "x-enum-varnames": [
                "OK",
                "OK_CREATE",
                "BAD_REQUEST",
                "UNAUTHORIZED",
                "NOT_FOUND",
                "METHOD_NOT_ALLOWED",
                "INTERNAL_SERVER_ERROR",
                "NOT_IMPLEMENTED",
                "BAD_GATEWAY",
                "GATEWAY_TIMEOUT",
                "HTTP_VERSION_NOT_SUPPORTED"
            ]
        }
    }
}